requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
PyMuPDF==1.24.10
Pillow==10.4.0
python-dotenv==1.0.1
openai==1.45.0
google-generativeai==0.8.3
gunicorn==23.0.0
asgiref==3.8.1

